+++ admin/compiler/joomla_3/HtmlView_edit.php
@@ -72,11 +72,11 @@ class ###Component###View###View### extends HtmlView
 			throw new Exception(implode("\n", $errors), 500);
 		}
 
-		// Display the template
-		parent::display($tpl);
-
 		// Set the document
 		$this->setDocument();
+
+		// Display the template
+		parent::display($tpl);
 	}