+++ admin/layouts/trashhelper.php
@@ -34,16 +34,16 @@ defined('JPATH_BASE') or die;
 				document.adminForm.elements['checkall-toggle'].checked=1;
 				Joomla.checkAll(document.adminForm.elements['checkall-toggle']);
 				// check to confirm the deletion
-				if(confirm('<?= Text::_("COM_COMPONENTBUILDER_ARE_YOU_SURE_YOU_WANT_TO_DELETE_CONFIRMING_WILL_PERMANENTLY_DELETE_THE_SELECTED_ITEMS") ?>')) {
-					Joomla.submitbutton('<?= $displayData->get("name") ?>.delete');
+				if(confirm('<?php echo Text::_("COM_COMPONENTBUILDER_ARE_YOU_SURE_YOU_WANT_TO_DELETE_CONFIRMING_WILL_PERMANENTLY_DELETE_THE_SELECTED_ITEMS") ?>')) {
+					Joomla.submitbutton('<?php echo $displayData->getName(); ?>.delete');
 				} else {
 					document.adminForm.elements['checkall-toggle'].checked=0;
 					Joomla.checkAll(document.adminForm.elements['checkall-toggle']);
 				}
 			} else {
 				// confirm deletion of those selected
-				if (confirm('<?= Text::_("COM_COMPONENTBUILDER_ARE_YOU_SURE_YOU_WANT_TO_DELETE_CONFIRMING_WILL_PERMANENTLY_DELETE_THE_SELECTED_ITEMS") ?>')) {
-					Joomla.submitbutton('<?= $displayData->get("name") ?>.delete');
+				if (confirm('<?php echo Text::_("COM_COMPONENTBUILDER_ARE_YOU_SURE_YOU_WANT_TO_DELETE_CONFIRMING_WILL_PERMANENTLY_DELETE_THE_SELECTED_ITEMS") ?>')) {
+					Joomla.submitbutton('<?php echo $displayData->getName(); ?>.delete');
 				};
 			}
 			return false;
@@ -59,23 +59,23 @@ defined('JPATH_BASE') or die;
 		<?php if (empty($displayData->items)): ?>
 			<h4 class="alert-heading">
 				<span class="icon-trash"></span>
-				<?= Text::_("COM_COMPONENTBUILDER_TRASH_AREA") ?>
+				<?php echo Text::_("COM_COMPONENTBUILDER_TRASH_AREA") ?>
 			</h4>
-			<p><?= Text::_("COM_COMPONENTBUILDER_YOU_ARE_CURRENTLY_VIEWING_THE_TRASH_AREA_AND_YOU_DONT_HAVE_ANY_ITEMS_IN_TRASH_AT_THE_MOMENT") ?></p>
+			<p><?php echo Text::_("COM_COMPONENTBUILDER_YOU_ARE_CURRENTLY_VIEWING_THE_TRASH_AREA_AND_YOU_DONT_HAVE_ANY_ITEMS_IN_TRASH_AT_THE_MOMENT") ?></p>
 		<?php else: ?>
 			<h4 class="alert-heading">
 				<span class="icon-trash"></span>
-				<?= Text::_("COM_COMPONENTBUILDER_TRASHED_ITEMS") ?>
+				<?php echo Text::_("COM_COMPONENTBUILDER_TRASHED_ITEMS") ?>
 			</h4>
-			<p><?= Text::_("COM_COMPONENTBUILDER_YOU_ARE_CURRENTLY_VIEWING_THE_TRASHED_ITEMS") ?></p>
+			<p><?php echo Text::_("COM_COMPONENTBUILDER_YOU_ARE_CURRENTLY_VIEWING_THE_TRASHED_ITEMS") ?></p>
 			<button onclick="emptyTrash();" class="btn btn-small btn-danger">
 				<span class="icon-delete" aria-hidden="true"></span>
-				<?= Text::_("COM_COMPONENTBUILDER_EMPTY_TRASH") ?>
+				<?php echo Text::_("COM_COMPONENTBUILDER_EMPTY_TRASH") ?>
 			</button>
 		<?php endif; ?>
 		<button onclick="exitTrash();" class="btn btn-small">
 			<span class="icon-back" aria-hidden="true"></span>
-			<?= Text::_("COM_COMPONENTBUILDER_EXIT_TRASH") ?>
+			<?php echo Text::_("COM_COMPONENTBUILDER_EXIT_TRASH") ?>
 		</button>
 	</div>
 <?php endif; ?>