+++ admin/src/View/Dynamic_gets/HtmlView.php
@@ -36,6 +36,7 @@ use VDM\Joomla\Utilities\StringHelper;
  *
  * @since  1.6
  */
+#[AllowDynamicProperties]
 class HtmlView extends BaseHtmlView
 {
 	/**